requirements.txt
Django==6.0.3
google-generativeai==0.8.6
python-dotenv==1.2.2
Pillow
